(Script to demonstrate callback retention)
This script produces a heap snapshot of the current process. Run the
script and open the file heapdump-*.heapsnapshot in the current
directory in the chrome profiler. Search for the class Marker. With the
`tryCatchTarget` patch applied there is only one of them (the prototype)
- without it there is the prototype and an instance retained by the
module variable `tryCatchTarget`.

The same case can be made for error objects.